requirements.txt
customtkinter==5.2.1
google_api_python_client==2.109.0
google_auth_oauthlib==1.1.0
Pillow==10.1.0
protobuf==4.25.1
py7zr==0.20.8
PyYAML==6.0.1
pyz3r==6.1.0
rarfile==4.1
SQLAlchemy==2.0.23
ttkbootstrap==1.10.1
